pyproject.toml
[project]
name = "ergo-type"
version = "0.1.0"
description = "Genetic Algorithm for optimizing keyboard layouts"
authors = [
    { name="Cătălin Pleșu", email="you@example.com" }
]
readme = "README.md"
requires-python = ">=3.11.0, <3.12"
dependencies = [
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.2",
    "typer>=0.9.0",
	"GitPython"
]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"] 

[tool.setuptools.packages.find]
include = [] 
